move arrow animation to Main.jsx
