+++ src/components/Main.jsx
@@ -1,4 +1,4 @@
-import React, { Component} from 'react';
+import React, { Component, createRef } from 'react';
 import { Jumbotron, Row, Col } from 'reactstrap';
 
 import '../assets/styles.css';
@@ -7,14 +7,32 @@ import Resume from '../resources/Anthony_Resume.pdf';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faFileAlt, faEnvelope } from '@fortawesome/free-solid-svg-icons'
 import { faGithub, faLinkedin } from '@fortawesome/free-brands-svg-icons'
+import { faAngleDoubleDown } from '@fortawesome/free-solid-svg-icons'
 
 class Main extends Component {
 
 	constructor (props){
 		super(props);
+		this.about = createRef();
+		this.state = { currOpacity: 1 };
+	}
+
+	componentDidMount () { 
+		window.onscroll =()=>{ 
+			if (window.scrollY < 750){
+				this.setState({currOpacity: 1})
+			}
+			if (750 <= window.scrollY){
+				const newOpacity = (200 - ((Math.ceil(window.scrollY/ 25)*25) - 750))/200;
+				this.setState({currOpacity: newOpacity})
+			}
+		};
+		console.log("did this log?")
+		// this.clickAbout.scrollIntoView();
 	}
 
 	render () {
+		const opacity = Math.min(this.state.currOpacity, 1);
 		return (
 		    <div className="Main">
 		    	<Jumbotron className="Main-jumbo">
@@ -39,6 +57,9 @@ class Main extends Component {
 							<FontAwesomeIcon className="Main-icons" icon={faEnvelope} size="3x" /></a></Col>
 					</Row>
 				</Jumbotron>
+		    	<span className="Main-arrow" style={{ opacity }}>
+					<FontAwesomeIcon className="Main-arrow-icon" icon={faAngleDoubleDown} size="3x" />
+				</span>
 		    </div>
 		)
 	}
